third_party/blink/web_tests/external/wpt/css/css-backgrounds/parsing/background-image-computed.sub-expected.txt

This is a testharness.js-based test.
[FAIL] Property background-image value 'radial-gradient(circle calc(-0.5em + 10px) at calc(-1em + 10px) calc(-2em + 10px), red, blue)'
  assert_equals: expected "radial-gradient(0px at -30px -70px, rgb(255, 0, 0), rgb(0, 0, 255))" but got "radial-gradient(calc(-0.5em + 10px) at calc(-1em + 10px) calc(-2em + 10px), rgb(255, 0, 0), rgb(0, 0, 255))"
[FAIL] Property background-image value 'radial-gradient(ellipse calc(-0.5em + 10px) calc(0.5em + 10px) at 20px 30px, red, blue)'
  assert_equals: expected "radial-gradient(0px 30px at 20px 30px, rgb(255, 0, 0), rgb(0, 0, 255))" but got "radial-gradient(calc(-0.5em + 10px) calc(0.5em + 10px) at 20px 30px, rgb(255, 0, 0), rgb(0, 0, 255))"
[FAIL] Property background-image value 'radial-gradient(ellipse calc(0.5em + 10px) calc(-0.5em + 10px) at 20px 30px, red, blue)'
  assert_equals: expected "radial-gradient(30px 0px at 20px 30px, rgb(255, 0, 0), rgb(0, 0, 255))" but got "radial-gradient(calc(0.5em + 10px) calc(-0.5em + 10px) at 20px 30px, rgb(255, 0, 0), rgb(0, 0, 255))"
[FAIL] Property background-image value 'conic-gradient(red 0deg, gold 1turn)'
  assert_equals: expected "conic-gradient(rgb(255, 0, 0) 0deg, rgb(255, 215, 0) 360deg)" but got "conic-gradient(rgb(255, 0, 0) 0deg, rgb(255, 215, 0) 1turn)"
Harness: the test ran to completion.
